0120 김수민 feat: 품목, 품목 공통 코드에서 이미 삭제된 상품은 버튼 안보이게, 백엔드에서도 검증

+++ backend/src/main/java/com/example/backend/controller/commonCode/CommonController.java
@@ -44,6 +44,13 @@ public class CommonController {
     @PutMapping("item/delete/{itemCommonCodeKey}")
     public ResponseEntity<Map<String, Object>> deleteItemCommonCode(
             @PathVariable int itemCommonCodeKey) {
+        // 이미 삭제된 품목 공통 코드인지 검증
+        if (service.deletedItemCommonCode(itemCommonCodeKey)) {
+            return ResponseEntity.badRequest().body(Map.of(
+                    "message", Map.of("type", "error", "text", "이미 삭제된 품목 공통 코드 입니다.")
+            ));
+        }
+
         if (service.deleteItemCommonCode(itemCommonCodeKey)) {
             return ResponseEntity.ok()
                     .body(Map.of("message", Map.of("type", "success",
